DRY & reorganization in general
