+++ components/About/About.tsx
@@ -4,7 +4,7 @@ import { View, ScrollView, SafeAreaView } from 'react-native';
 import { useTheme } from '@react-navigation/native';
 
 import FadeText from '../Components/FadeText';
-import Button from '../Button';
+import Button from '../Components/Button';
 import { ThemeType } from '../../app/types';
 import { ContextAppLoaded } from '../../app/context';
 import Header from '../Header';
