--- components/CircularProgress/index.js
@@ -1,3 +1,0 @@
-import CircularProgress from './CircularProgress';
-
-export default CircularProgress;
