+++ components/Components/DetailLine.tsx
@@ -3,9 +3,9 @@ import React, { ReactNode } from 'react';
 import { View } from 'react-native';
 import { useTheme } from '@react-navigation/native';
 
-import FadeText from '../../Components/FadeText';
-import RegText from '../../Components/RegText';
-import { ThemeType } from '../../../app/types';
+import FadeText from './FadeText';
+import RegText from './RegText';
+import { ThemeType } from '../../app/types';
 
 type DetailLineProps = {
   label: string;
@@ -16,7 +16,7 @@ type DetailLineProps = {
 const DetailLine: React.FunctionComponent<DetailLineProps> = ({ label, value, children }) => {
   const { colors } = useTheme() as unknown as ThemeType;
   return (
-    <View style={{ display: 'flex', marginTop: 20, flexDirection: 'row', alignItems: 'center' }}>
+    <View style={{ display: 'flex', marginTop: 20 }}>
       <FadeText>{label}</FadeText>
       <View style={{ width: 10 }} />
       {value && <RegText color={colors.text}>{value}</RegText>}
