+++ components/ImportKey/components/ScannerKey.tsx
@@ -5,16 +5,16 @@ import { useTheme } from '@react-navigation/native';
 import QRCodeScanner from 'react-native-qrcode-scanner';
 
 import RegText from '../../Components/RegText';
-import Button from '../../Button';
+import Button from '../../Components/Button';
 import { useContext } from 'react';
 import { ContextAppLoaded } from '../../../app/context';
 import { BarCodeReadEvent } from 'react-native-camera';
 
-type ScannerProps = {
+type ScannerKeyProps = {
   setPrivKeyText: (k: string) => void;
   closeModal: () => void;
 };
-const Scanner: React.FunctionComponent<ScannerProps> = ({ setPrivKeyText, closeModal }) => {
+const ScannerKey: React.FunctionComponent<ScannerKeyProps> = ({ setPrivKeyText, closeModal }) => {
   const context = useContext(ContextAppLoaded);
   const { translate } = context;
   const validateKey = (scannedKey: string) => {
@@ -57,4 +57,4 @@ const Scanner: React.FunctionComponent<ScannerProps> = ({ setPrivKeyText, closeM
   );
 };
 
-export default Scanner;
+export default ScannerKey;
