+++ components/Send/Send.tsx
@@ -13,12 +13,12 @@ import ErrorText from '../Components/ErrorText';
 import RegText from '../Components/RegText';
 import ZecAmount from '../Components/ZecAmount';
 import CurrencyAmount from '../Components/CurrencyAmount';
-import Button from '../Button';
+import Button from '../Components/Button';
 import { SendPageStateClass, SendProgressClass, ToAddrClass } from '../../app/AppState';
 import { parseZcashURI, ZcashURITarget } from '../../app/uris';
-import RPCModule from '../RPCModule';
+import RPCModule from '../../app/RPCModule';
 import Utils from '../../app/utils';
-import Scanner from './components/Scanner';
+import ScannerAddress from './components/ScannerAddress';
 import Confirm from './components/Confirm';
 import { ThemeType } from '../../app/types';
 import { ContextAppLoaded } from '../../app/context';
@@ -360,7 +360,7 @@ const Send: React.FunctionComponent<SendProps> = ({
         transparent={false}
         visible={qrcodeModalVisble}
         onRequestClose={() => setQrcodeModalVisible(false)}>
-        <Scanner
+        <ScannerAddress
           updateToField={updateToField}
           closeModal={() => setQrcodeModalVisible(false)}
           width={dimensions.width - 42}
@@ -760,7 +760,7 @@ const Send: React.FunctionComponent<SendProps> = ({
           transparent={false}
           visible={qrcodeModalVisble}
           onRequestClose={() => setQrcodeModalVisible(false)}>
-          <Scanner
+          <ScannerAddress
             updateToField={updateToField}
             closeModal={() => setQrcodeModalVisible(false)}
             width={dimensions.width / 2}
